# Speedup Element::floordiv