--- a/news/floordiv.rst
+++ b/news/floordiv.rst
@@ -1,0 +1,3 @@
+**Changed:**
+
+* speed up floor division